fix(test): move InitTestMarshal test behind testtools tag
Keeps the init marshaler coverage test in a testtools-only file so CI
lint and compile succeed when cross-referencing test helpers.
